Fixed stats collection instructions
The old version of the read me suggested that you should use the "word-stats" npm package.  However, that package has now has either been change or moved.  If you run the code in the old read me, it tells you that word-stats does not have a function calculate.  In addition, the word-stats package does not produce the markov matrices that we need any way.  It only outputs statistics on individuals words pronunciation, syllable structure and emphasis.  The read me now tells you to use the calculateStats function from this package which does what you need.